pyproject.toml
[tool.poetry]
name = "podcast-transcriber"
version = "0.1.0"
description = ""
authors = ["Ashley Felton <ashley@ropable.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
feedparser = "6.0.11"
requests = "^2.31.0"
openai-whisper = "^20231117"
python-slugify = "^8.0.3"

[tool.poetry.group.dev.dependencies]
ipython = "^8.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
